Fix major ACF conflict. Fix PHP notice.
fixes #108, fixes #106, fixes #105, resolves #100

issue 100 is resolved because it was never actually the problem. The ACF issue was the problem.

The ACF issue was fixed by making sure that the editor used was always verified to have an ID of "content". ACF added a secondary, hidden editor to post edit pages, and Render was targeting this editor instead of the main editor.